Move methods from discovery.utils.filesystem.base to filesystem module.
